requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
boto3==1.20.46
botocore==1.23.46
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-allauth==0.47.0
django-heroku==0.3.1
django-location-field==2.1.0
django-sendgrid-v5==1.2.0
django-smtp-ssl==1.0
django-storages==1.12.3
docopt==0.6.2
Flask==2.0.2
google==3.0.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud==0.34.0
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
httplib2==0.20.2
httpserver==1.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
pep8==1.7.1
Pillow==9.0.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
schedule==1.1.0
sendgrid==6.9.7
sentry-sdk==1.5.12
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
whitenoise==5.3.0